Add files via upload
The automated tools you use on Debian-based systems often tell you which directory and directory path it is.

However, you need to parse these files most of the time.
There are some files that forensic analysts prioritize on debian-based systems without knowing the details of the case. I wrote a script that parses them for the convenience of the analysts. I wish you pleasant legal events. =)